updated UPC and changed to the correct datatypes inside DatabaseHandler: #61 #34

+++ app/src/main/java/com/example/josegeorges/paintit/POJO/Brushes.java
@@ -5,7 +5,7 @@ public class Brushes extends Item {
     private int brushImageView;
     private String brushName;
 
-    public Brushes(int upc, double price, String description, int imageView, int size, int itemTypeId, String paintName) {
+    public Brushes(long upc, double price, String description, int imageView, int size, int itemTypeId, String paintName) {
         super(upc, price, description, imageView, size, itemTypeId);
         this.brushName = paintName;
     }
